requirements.txt
streamlit==1.28.0
langchain
langchain-community
langchain-huggingface
sentence-transformers>=2.6.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
transformers>=4.39.0
torch
accelerate
bitsandbytes
